+++ config/filament-latex.php
@@ -18,6 +18,8 @@ return [
      */
     'user-model' => 'App\Models\User',
     'storage' => 'private',  // If you want to change the storage, you have to create a filesystem disk in config/filesystems.php
+    'storage-url' => '/private_storage', // The URL to the storage disk. This is used to generate the download link.
     'parser' => '/usr/bin/pdflatex', // The latex parser to use. Options: pdflatex, xelatex, lualatex (pdflatex is the default).
     'compilation-timeout' => 60, // The maximum time in seconds to wait for the compilation to finish.
+    'strict-compilation' => false, // Options: strict, non-strict. Strict mode will throw exceptions on compilation errors, otherwise it will not.
 ];