+++ resources/views/page.blade.php
@@ -1,5 +1,6 @@
 @php
-$latexContent = $this->latexContent;
+    $latexContent = $this->latexContent;
+    $pdfUrl = $this->getPdfUrl();
 @endphp
 
 <x-filament-panels::page>
@@ -25,7 +26,7 @@ $latexContent = $this->latexContent;
             >
                 {{-- Latex Editor --}}
                 <div
-                    class="w-full border border-gray-200 rounded-lg dark:border-gray-700"
+                    class="w-full border border-gray-200 rounded-lg dark:border-gray-700 h-[75rem] overflow-auto"
                     x-ignore
                     ax-load
                     x-model="message"
@@ -37,9 +38,21 @@ $latexContent = $this->latexContent;
                 >
                 </div>
 
-                {{-- Latex Preview --}}
+                {{-- PDF Preview --}}
                 <div
-                    class="border border-gray-200 rounded-lg dark:border-gray-700">
+                    class="border border-gray-200 rounded-lg dark:border-gray-700"
+                >
+                    @if($pdfUrl)
+                        <iframe
+                            x-data="{ pdfUrl: @js($pdfUrl) }"
+                            x-on:document-compiled.window="pdfUrl = @js($pdfUrl) + '?' + new Date().getTime()"
+                            class="w-full h-[75rem]"
+                            :src="pdfUrl"
+                        >
+                        </iframe>
+                    @else
+                        <p>No PDF available to preview.</p>
+                    @endif
                 </div>
             </div>
         </x-filament::section>